refactor: split local_input_actuator.py into 4 sub-modules
Per galgame-pr4-split-plan step 5. Pure cut/paste — every function
body and module-level constant is byte-identical to HEAD (verified
via AST diff on all 40 defs and 27 assigns).

Splits:
- _win32_input_types.py (23 lines): Win32 numeric constants (VK_*,
  KEYEVENTF_*, MOUSEEVENTF_*, SW_RESTORE, INPUT_*, MAPVK_VK_TO_VSC,
  PROCESS_QUERY_LIMITED_INFORMATION, TOKEN_QUERY, TokenElevation).
- _window_manager.py (355 lines): ctypes structures (TOKEN_ELEVATION,
  RECT, MOUSEINPUT, KEYBDINPUT, INPUT_UNION, INPUT), module state
  (_LAST_FOCUS_WINDOW_DIAGNOSTIC*, _WAIT_EVENT, _LOGGER), focus
  diagnostic helpers, _runtime_target, _find_window_for_pid,
  _window_text, _root/_window_process_id, _focus_window, elevation
  checks.
- _input_primitives.py (221 lines): _matching_input_safety_deny_marker,
  _input_safety_policy_block_reason, _tap_key, _click,
  _client_screen_rect, _rect_payload, _coerce_rect/_coerce_source_size,
  _relative_point_forbidden_zone. Owns INPUT_SAFETY_DENY_MARKERS and
  VIRTUAL_MOUSE_FORBIDDEN_ZONES (moved from main since used only here).
- _choice_resolver.py (236 lines): _snapshot_*, _choose_*,
  _resolve_virtual_mouse_dialogue_target,
  _resolve_choice_bounds_click_target, _recover_should_press_escape,
  _looks_like_system_menu. Owns SYSTEM_MENU_MARKERS and
  VIRTUAL_MOUSE_DIALOGUE_CANDIDATES (moved from main).
- local_input_actuator.py (292 lines, was 994): slim shim with
  perform_local_input_actuation, try_focus_target_window plus
  re-exports of all 67 public/internal names for backward compat.

GitNexus reports MEDIUM risk (27 symbols touched, 1 affected process,
1 changed step) — purely relocation noise; logic is byte-identical.
