+++ plugin/plugins/galgame_plugin/_win32_input_types.py
@@ -1,0 +1,23 @@
+from __future__ import annotations
+
+
+VK_RETURN = 0x0D
+VK_ESCAPE = 0x1B
+VK_SPACE = 0x20
+VK_UP = 0x26
+VK_DOWN = 0x28
+
+KEYEVENTF_KEYUP = 0x0002
+KEYEVENTF_SCANCODE = 0x0008
+KEYEVENTF_EXTENDEDKEY = 0x0001
+MOUSEEVENTF_LEFTDOWN = 0x0002
+MOUSEEVENTF_LEFTUP = 0x0004
+MOUSEEVENTF_MOVE = 0x0001
+MOUSEEVENTF_ABSOLUTE = 0x8000
+SW_RESTORE = 9
+INPUT_MOUSE = 0
+INPUT_KEYBOARD = 1
+MAPVK_VK_TO_VSC = 0
+PROCESS_QUERY_LIMITED_INFORMATION = 0x1000
+TOKEN_QUERY = 0x0008
+TokenElevation = 20
